KPLN_BIMTools_Ribbon. BatchAddingParam - добавлена функция внесения значений параметров (всех*)
