+++ KPLN_BIMTools_Ribbon/KPLN_BIMTools_Ribbon/Forms/batchAddingParametersWindowСhoice.xaml
@@ -11,7 +11,7 @@
         Topmost="True" 
         WindowStartupLocation="CenterScreen" 
         Width="800" 
-        Height="210" 
+        Height="240" 
         ResizeMode="NoResize"
         Background="#FF292D36">
     <Grid>
@@ -19,7 +19,8 @@
         <StackPanel Orientation="Vertical" VerticalAlignment="Center" Margin="0,30,0,0">
             <Button Content="Пакетное добавление общих параметров" Height="25px" Width="600px" Margin="0,0,0,5" Click="Button_NewGeneralParam"/>
             <Button Content="Пакетное добавление параметров семейства" Height="25px" Width="600px" Margin="0,0,0,5" Click="Button_NewFamilyParam"/>
-            <Button Content="Использовать готовый файл-конфигурации" Height="25px" Width="600px" Click="Button_LoadParam" Background="#FFBAFDAD"/>
+            <Button Content="Использовать готовый файл-конфигурации" Height="25px" Width="600px" Margin="0,0,0,5" Click="Button_LoadParam" Background="#FFBAFDAD"/>
+            <Button Content="Использовать готовый файл-конфигурации для нескольких семейств" Height="25px" Width="600px" Background="#FF6DBF5D"/>
         </StackPanel>
     </Grid>
 </Window>
